Release v1.12.0
This release introduces a new feature for the crash probe. It will now
collect source file and line number information for each frame in a
backtrace if the data is available.

Also, this release saw the beginning of migrating the code base to use
libnica instead of glib, with the goal of reducing memory footprint of
the telemetry daemon and probes. Future releases will convert more
existing glib usage to libnica.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>