Properly init/deinit environment for libcurl
The libcurl documentation states that curl_global_init() "must" be
called within a program that uses the library, but because the libcurl
integration code has been working for a long time, this is clearly not
an absolute requirement.

We did however detect a memory resource leak that is fixed by adding
calls to curl_global_init() and curl_global_cleanup(), so add the
appropriate calls to fix the issue.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>