Revert "Removing crash dump configuration"
Installing a crashprobe sysctl conf file by default is, in my opinion,
the most useful way to enable the crashprobe for Linux distributions.

The fact that Clear Linux does not use it anymore does not necessarily
dictate policy for telemetrics-client itself. My preference would be to
simply exclude the conf file in the Clear Linux telemetrics-client
package, since that is very easy to do with autospec. Also, the new
crash handling implementation in Clear Linux requires a systemd patch
that won't be upstreamed, so this is another good reason to install the
conf by default.

This reverts commit d1425173a9fa29a643c2eb0a32a7b73869dcf978.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>