Spool: retry more often if there's work to do
The spool error retry time is set to 900s by default, which
means that in any spool error case, the system will wait 15
minutes before trying to deliver a single record again.

This causes us to lose a lot of telemetry because (1) telemetry
ends up starting before networking is active and (2) the system
shuts down after a few minutes already, never making it to the
full 15 minute mark.

This change adds a throttle that automatically extends the time
between delivery attempts when the spool isn't empty and a
delivery error occurs. It simply retries the delivery after 1, 2,
4, 8, etc. seconds up to 256 until the spool either delivers all
the entries, or until the throttle end.

This makes the 900s retry time mostly useless, in effect.