python-probe: Initial commit
The purpose of the python-probe is to monitor a specific folder
(as defined by PYTHON_TELEMETRY_DIR) for any Python generated reports.
If any such file is detected, it is sent to the telemetry backend
for analysis.

The reports are typically tracebacks of unhandled Python exceptions
caused by Python programs residing in /usr/bin. It is the responsibility
of Python to generate the traceback files and place them in
PYTHON_TELEMTRY_DIR. The probe will only detect the presence of the files
and send the content of the files as a payload.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>