tm_set_config_file: return status and call from all probes
This change was on a TODO list for a few years.
The function tm_set_config_file could silently fail as it did not
return any status. This forced implementing of all error checks by
the callers of this function.
The current implementation does its own error checks and string copying.
The function returns an error code if errors were encountered.
Since this is a library function, all callers needed to be modified as well.
However, the modifications are code simplifications so this is not too painful.
In addition, the change also affected library documentation as well (void -> int).

As a direct consequence, this change allows simple implementation of the command
line argument "--config_file" uniformly across all probes.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>