telempostdaemon: fix retries loop
The 'else if (retry_attempt == MAX_RETRY_ATTEMPTS)' section would get
hit the first time through the loop on every invokation, printing the
failed to deliver error to the journal.

To fix it, initialize retry_attempt to zero and check that we're above
zero retries to enter the retry section.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>