telemetry.c: support variants in the system_name header
Read the VARIANT_ID field of the os-release file and add it to our
system_name header. This allows differentiation of different variants of
an operating system on the backend.

The header will be unchanged on operating systems without a VARIANT_ID
in their os-release file. Also continue limiting the header length to 80
characters.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>