src: standardize log and debug messages
Use the telem_log, telem_debug, and telem_perror macros to print
messages instead of printf and fprintf where appropriate. These macros
may be configured to print to many standard locations at compile time,
such as the syslog, the journal, or stderr. The remaining uses of
fprintf in printing errors is only for command line utilities.

When using telem_debug, we don't need to surround it with #ifdef DEBUG
statements, as the macro does that itself.

Finally, stop hiding real errors behind #ifdef DEBUGs. If an error
occurs, it should be logged.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>